subroutine of steadystate.m
Matlab code called directly while running steadystate.m. The code allows the user to select the 4 points required to run steadystate and to compare the slopes before and after the velocity steps (i.e., start point of the analysis before the velocity step, point corresponding to the velocity step, first point to calculate the slope after the velocitry step and last point of the analysis).