Main code
Matlab code used to find the optimum pair of first steady-state point after a velocity step and its associated slip window starting the steady-state point. The goal of the program is to allow the user to operate a sensible detrend in order to determine accurate rate- and state- friction parameter values from numerical inversions.